update: passed user id across components
