+++ src/app/products/[productId]/page.tsx
@@ -2,8 +2,14 @@ import ProductDetailsPage from "@/components/productDetails/ProductDetailsPage";
 import { getQueryClient } from "@/providers/get-query-client";
 import { getSingleProduct } from "@/utils/api/api";
 import { dehydrate, HydrationBoundary, useQuery } from "@tanstack/react-query";
+import { auth } from "../../../../auth";
 
-export default function ProductIdHome({ params }: { params: { productId: string } }) {
+export default async function ProductIdHome({
+  params,
+}: {
+  params: { productId: string };
+}) {
+  const session = await auth();
   const queryClient = getQueryClient();
 
   queryClient.prefetchQuery({
@@ -14,7 +20,7 @@ export default function ProductIdHome({ params }: { params: { productId: string 
   return (
     <section className="flex min-h-[calc(100vh-96px)] w-full mx-auto flex-col mt-24 md:mt-16 ">
       <HydrationBoundary state={dehydrate(queryClient)}>
-        <ProductDetailsPage productId={params.productId} />
+        <ProductDetailsPage userId={session?.user?.id} productId={params.productId} />
       </HydrationBoundary>
     </section>
   );
