+++ src/components/account/wishlist/delete-product.tsx
@@ -1,16 +1,20 @@
-"use client";
-
+import { Button } from "@/components/ui/button";
 import { removeFromWishlist } from "@/lib/actions";
 import { Trash } from "lucide-react";
-import { useSession } from "next-auth/react";
 import { useRouter } from "next/navigation";
 import React from "react";
 import { toast } from "sonner";
 
-const DeleteProduct = ({ productId, refetch }: { productId: number, refetch: () => void }) => {
+const DeleteProduct = ({
+  userId,
+  productId,
+  refetch,
+}: {
+  userId: string | undefined;
+  productId: number;
+  refetch: () => void;
+}) => {
   const router = useRouter();
-  const { data: session } = useSession();
-  const userId = session?.user?.id;
 
   const handleDeleteWishlist = async () => {
     if (!userId) {
@@ -32,12 +36,16 @@ const DeleteProduct = ({ productId, refetch }: { productId: number, refetch: () 
 
   return (
     <div>
-      <Trash
-        className="absolute h-4 w-5 fill-gray-200 stroke-slate-400 top-2 right-2 hover:cursor-pointer"
+      <Button
+        variant={"ghost"}
+        size={"icon"}
+        className="absolute top-2 right-2 group"
         onClick={() => {
           handleDeleteWishlist();
         }}
-      />
+      >
+        <Trash className=" h-4 w-5 fill-gray-200 stroke-slate-400 group-hover:fill-red-400 group-hover:stroke-red-700" />
+      </Button>
     </div>
   );
 };
